Extracted project name de-duplication into a service
