+++ subprojects/ide/src/main/java/org/gradle/plugins/ide/internal/tooling/ToolingModelServices.java
@@ -17,10 +17,14 @@
 package org.gradle.plugins.ide.internal.tooling;
 
 import org.gradle.api.internal.artifacts.ivyservice.projectmodule.ProjectPublicationRegistry;
+import org.gradle.api.internal.project.ProjectIdentifier;
+import org.gradle.api.internal.project.ProjectRegistry;
 import org.gradle.api.internal.project.ProjectTaskLister;
 import org.gradle.internal.service.ServiceRegistration;
 import org.gradle.internal.service.ServiceRegistry;
 import org.gradle.internal.service.scopes.PluginServiceRegistry;
+import org.gradle.plugins.ide.internal.configurer.DefaultUniqueProjectNameProvider;
+import org.gradle.plugins.ide.internal.configurer.UniqueProjectNameProvider;
 import org.gradle.tooling.provider.model.ToolingModelBuilderRegistry;
 import org.gradle.tooling.provider.model.internal.DefaultToolingModelBuilderRegistry;
 
@@ -47,6 +51,10 @@ public class ToolingModelServices implements PluginServiceRegistry {
     }
 
     private static class BuildScopeToolingServices {
+
+        protected UniqueProjectNameProvider createBuildProjectRegistry(ProjectRegistry<ProjectIdentifier> projectRegistry) {
+            return new DefaultUniqueProjectNameProvider(projectRegistry);
+        }
 
         protected ToolingModelBuilderRegistry createBuildScopedToolingModelBuilders(ProjectTaskLister taskLister, ProjectPublicationRegistry projectPublicationRegistry, ServiceRegistry services) {
             DefaultToolingModelBuilderRegistry registry = new DefaultToolingModelBuilderRegistry();
